# added binary version of vocabulary to speedup ORB_SLAM2 start