I am a noob at both C++ and git, so please be forgiving.

This is a patch to speedup ORB-SLAM2 startup by loading the vocabulary from a binary file instead of ascii.

I tested it on linux on an intel I7 and an arm Exynos5422 (odroid xu4)
          ascii   | binary
I7     12.06s | 0.2s
arm  77s      | 1s

(ORB_SLAM2 was just not usable on the arm with a startup duration of 77s)

I added a line to build.sh to convert  Vocabulary/ORBvoc.txt to Vocabulary/ORBvoc.bin

ORB_SLAM2 uses file extension to choose between ascii and binary.

Test with
./Examples/Monocular/mono_tum Vocabulary/ORBvoc.bin Examples/Monocular/TUM1.yaml /data/tum/rgbd_dataset_freiburg1_xyz/
versus
./Examples/Monocular/mono_tum Vocabulary/ORBvoc.txt Examples/Monocular/TUM1.yaml /data/tum/rgbd_dataset_freiburg1_xyz/

Regards

Poine