@hashten 2016-09-16
This works extremely nice straight out of the box! From a loading time of 20.8 sec to 0.48 sec on my 5 year old really slow AMD A6-3410MX. Thanks a bunch, makes testing much more fun!