@poine 2016-09-17
You're very welcome. You make the that time i spend preparing the patch
worth.

On Fri, Sep 16, 2016 at 6:07 PM, Joel Harsten <notifications@github.com>
wrote:

> *@hashten* approved this pull request.
>
> Works straight away and changed loading time from 20 sec to 0.5 sec!
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/raulmur/ORB_SLAM2/pull/21#pullrequestreview-381664>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAGTzzNueT8k7M-DphLGWRcIImzEeADRks5qqr7DgaJpZM4HX6iC>
> .
>