@tdnet12434 2016-10-06
Hi @poine ,
Thx u for your patch, this can save a lot of time for doing my thesis!!
I wanna ask a question. How smooth that ORB_SLAM2 running on odroid XU4? I think i will buy one but i not sure about its max frame rate handling.