@poine 2016-10-06
@Thanabadee Blsc

It is difficult to answer your question. Framerate depends on the
resolution of your video, the number of extracted features and so on. But
all in all, i don't think you'll get a great framerate with an XU4 (like
maybe a few fps).

If you are concerned about miniaturization and power consumption, i would
advise to take a look at nvidia tegra based boards as you might be able to
take profit of the cuda cores. Parrot has recently released a tegra based
dev board that look very nice:
https://corporate.parrot.com/en/pressrelease/parrots.l.a.m.dunkturnadroneintoasmartrobot

If power consumption is not critical to your application, i would suggest
to make your life easier and use the fastest intel processor you can
afford. This is the path i am following. This was my first robot, with the
XU4: https://goo.gl/photos/WhxoUfWTV6JWFAtn8
This is my current robot: https://goo.gl/photos/17ptWpjAzf2B6gfK7
I am just using an I7 powered laptop on it. I do not regret switching to
the larger robot, which by the way, was not that much more expensive than
the small one (it's an old electric wheelchair). I would call my first
experiment with the small robot premature optimization.

Hth

Poine


On Thu, Oct 6, 2016 at 6:52 PM, Thanabadee Blsc <notifications@github.com>
wrote:

> Hi @poine <https://github.com/poine> ,
> Thx u for your patch, this can save a lot of time for doing my thesis!!
> I wanna ask a question. How smooth that ORB_SLAM2 running on odroid XU4? I
> think i will buy one but i not sure about its max frame rate handling.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/raulmur/ORB_SLAM2/pull/21#issuecomment-252022685>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAGTz9s8EKTszJyXd5J30IN-Jes-NHWRks5qxSdCgaJpZM4HX6iC>
> .
>