@sunghoon031 2016-10-21
@poine  Did you try with localization-only mode on XU4? Or was it with mapping + loop-closing?