@busybyte1 2018-06-19
For Windows Visual Studio 2017 :
Compiler-Error like: Not interpreted as constant

Thirdparty/DBoW2/DBoW2/TemplatedVocabulary.h
Line 1484:
-char buf[size_node];
+char *buf = new char[size_node]; 

rest works perfect, many thanks ! (0.45sec to load)