@busybyte1 2018-07-03
if you managed to get orb to work you should know what about i write next.

1. open dbow.sln project, which you should have build with cmake, with VisualStudio . 
2. in your project-explorer go to external dependancies and open TemplateVocabulary.h.
3. change the lines like in (https://github.com/raulmur/ORB_SLAM2/commit/4122702ced85b20bd458d0e74624b9610c19f8cc) and recompile 
4. same do for your orb-slam2 project, wich should be build with cmake, for system.cc and bin_vocabulary.cc
5. read the code to change intensive, theres a function called "save_as_binary". with this function you can save the vocabulary as bin
6. change the voc-filename in your mono/rgbd_tum from .txt to .bin