@Fourierw 2023-06-25
After adding the code I completely rebuilt the 'ORB_SLAM2' project , yet the **ORBvoc.bin** was nowhere to find.
Then I debugged the 'bin_vocabulary' project , it said **"Vocabulary loading failure: This is not a correct text file!"** , how can I solve this to get the binary file of ORBvoc ?
(when I run the mono_tum project , the ORB.txt file can be read correctly and the entire program was fine)