@jeremysalwen 2024-07-26
Disclaimer: I don't intend to create spam. If the ORB_SLAM2 developers wish me to stop posting, I will cease immediately.

ORB_SLAM2 seems to be inactive, but there are a number of community patches, including yours, which provide important bugfixes or improvements. Therefore I have created a new project ORB_SLAM_COMMUNITY (https://github.com/jeremysalwen/ORB_SLAM_COMMUNITY) to gather together these improvements and continue active development. I have already merged this PR in the ORB_SLAM_COMMUNITY version. We would welcome any other contributions you might have.